Update drag style

+++ services/native/src/devicestatus_service.cpp
@@ -763,7 +763,7 @@ int32_t DeviceStatusService::GetShadowOffset(int32_t& offsetX, int32_t& offsetY,
 int32_t DeviceStatusService::UpdateDragStyle(DragCursorStyle style)
 {
     CALL_DEBUG_ENTER;
-    int32_t tid = GetCallingTokenID();
+    int32_t tid = static_cast<int32_t>(GetCallingTokenID());
     int32_t pid = GetCallingPid();
     int32_t ret = delegateTasks_.PostSyncTask(
         std::bind(&DragManager::UpdateDragStyle, &dragMgr_, style, pid, tid));
